ML Model/requirements.txt
Flask==3.0.2
pandas==2.0.3
numpy==1.26.3
matplotlib==3.8.2
gunicorn==21.2.0
tensorflow==2.15.0
keras==2.15.0
pillow==10.0.1
